Refactor React Event name
