+++ src/imio/smartweb/core/webcomponents/src/components/Events/EventCard/EventCard.jsx
