+++ src/imio/smartweb/core/webcomponents/src/components/Events/EventContent/EventContent.jsx
